+++ example/lib/echo/socket_io.dart
@@ -4,13 +4,12 @@ import 'package:socket_io_client/socket_io_client.dart';
 const String token = 'API_TOKEN';
 
 Echo<Socket, SocketIoChannel> initSocketIOClient() => Echo.socket(
-      'http://localhost:3000',
+      'http://localhost:6001',
       autoConnect: false,
       auth: {
         'headers': {'Authorization': 'Bearer $token'}
       },
       moreOptions: {
-        // 'transports': ['websocket', 'polling', 'flashsocket']
         'transports': ['websocket']
       },
     );